refactor: Repository를 infrastructure 레이어로 이동
- domain 패키지의 JpaRepository를 infrastructure/persistence로 이동
- Clean Architecture 의존성 방향 준수
- JpaConfig에서 domain 패키지 스캔 제거
