+++ apps/commerce-streamer/src/main/java/com/loopers/application/metrics/MetricsAggregationService.java
@@ -5,9 +5,9 @@ import com.loopers.application.event.like.ProductUnlikedEvent;
 import com.loopers.application.event.order.OrderCompletedEvent;
 import com.loopers.application.event.product.ProductViewedEvent;
 import com.loopers.domain.event.EventHandled;
-import com.loopers.domain.event.EventHandledRepository;
 import com.loopers.domain.metrics.ProductMetrics;
-import com.loopers.domain.metrics.ProductMetricsRepository;
+import com.loopers.infrastructure.persistence.EventHandledRepository;
+import com.loopers.infrastructure.persistence.ProductMetricsRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
