+++ apps/commerce-streamer/src/test/java/com/loopers/application/metrics/MetricsAggregationServiceIdempotencyTest.java
@@ -5,9 +5,9 @@ import com.loopers.application.event.like.ProductUnlikedEvent;
 import com.loopers.application.event.order.OrderCompletedEvent;
 import com.loopers.application.event.product.ProductViewedEvent;
 import com.loopers.config.TestConfig;
-import com.loopers.domain.event.EventHandledRepository;
 import com.loopers.domain.metrics.ProductMetrics;
-import com.loopers.domain.metrics.ProductMetricsRepository;
+import com.loopers.infrastructure.persistence.EventHandledRepository;
+import com.loopers.infrastructure.persistence.ProductMetricsRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
